Release v0.1.0 fixes
* chore: add Lombok library

* feat: add main controllers

* config: set port to 8080

* fix: correct GetMapping paths

* feat: add cards page

* feat: add cardService

* feat: add CardRepository

* feat: add properties for db connection

* feat: add card info page

* feat: modified controller

* feat: modified models

* feat: modified card page

* REB-3, REB-4, REB-5, REB-6 Base structure (#4)

* feat: add REST controller

* feat: correct components for the Card

* feat: add components for the Note entity

* feat: add components for the Set entity

* feat: add updateCard dto

* feat: add updateNote dto

* feat: add HTTP statuses 201 and 204 for responses

* refactor: rename the model components to Dto

* fix: changed the field injection to constructor injection

* fix: added one common type for ObjectNotFound Exceptions

* feat: Add Mappers for converting between Dto and Entity

* feat: DTO as an returned object instead of Card in controllers

* refactor: changed the descr to description

* feat: ObjectNotFoundException extends RuntimeException

* refactor: update methods in services

* Feature/reb 7 improve exception handling (#5)

* feat: ControllerAdvice components added

* feat: redesigned Card components to work with ControllerAdvice

* feat: added internal server error handling

* REB-8 Introduce version control for db (#6)

* chore: added FlyWay

* feat: added tables

* REB-9 add integration tests (#7)

* feat: added card test

* refactor: test properties

* refactor: moved test properties to new configuration

* feat: tests do not share a single instance of test data

* feat: added cards creation in tests via repository call

* fix: changed the flyway dependency

* fix: typo in naming

* fix: db properties

* refactor: application properties

* fix: h2 dependency

* refactor: CardControllerTest method names

* refactor: CardControllerTest method names

* fix: db properties

* REB-10 add service level logs (#9)

* REB 9 add integration tests (#8)

* feat: added card test

* refactor: test properties

* refactor: moved test properties to new configuration

* feat: tests do not share a single instance of test data

* feat: added cards creation in tests via repository call

* fix: changed the flyway dependency

* chore: add logback

* refactor: log properties

* feat: added logs

* fix: db properties

* refactor: delete aplication-test.properties

* refactor: delete application.properties changes

* refactor: application.properties

* refactor: application.properties

* REB-11 add user entity (#10)

* REB 9 add integration tests (#8)

* feat: added card test

* refactor: test properties

* refactor: moved test properties to new configuration

* feat: tests do not share a single instance of test data

* feat: added cards creation in tests via repository call

* fix: changed the flyway dependency

* chore: add logback

* refactor: log properties

* feat: added logs

* fix: db properties

* refactor: delete aplication-test.properties

* refactor: delete application.properties changes

* refactor: application.properties

* refactor: application.properties

* feat: added user entity

* feat: added user to db migration

* feat: added user_id to entities and dto

* feat: added user components

* feat: added ThreadLocalUserIdHolder

* feat: added user id assignment

* fix: V3 migration

* fix: added lowerCamelCase names associated with user

* feat: added table existence check to migrations

* fix: added lowerCamelCase to UserMapper

* REB-17 App doesn't run migrations on startup (#15)

* fix: flyway migrations

* fix: application properties

* REB-18 Tests use non-test DB and fail (#16)

* refactor: deleted SmokeTest

* fix: h2 dependency

* fix: application properties (#17)

---------

+++ src/main/java/com/example/rebrain/controllers/CardController.java
@@ -28,7 +28,7 @@ public class CardController {
     public ResponseEntity<CardDto> createCard(@RequestBody CardPostDto cardDto) {
         log.debug("Creating card with data: {}", cardDto);
         CardEntity cardEntity = CardMapper.toEntity(cardDto.getCard());
-        Integer setId = cardDto.getSetId();
+        Long setId = cardDto.getSetId();
         CardEntity createdCard = cardService.create(cardEntity, setId);
         CardDto createdToDto = CardMapper.toDto(createdCard);
         URI location = URI.create("/cards/" + createdToDto.getId());
@@ -48,7 +48,7 @@ public class CardController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<CardDto> getCardById(@PathVariable Integer id) {
+    public ResponseEntity<CardDto> getCardById(@PathVariable Long id) {
         log.debug("Fetching card with ID: {}", id);
         CardDto card = CardMapper.toDto(cardService.getOne(id));
         log.debug("Card found: {}", card);
@@ -56,7 +56,7 @@ public class CardController {
     }
 
     @PatchMapping("/{id}")
-    public ResponseEntity<CardDto> updateCard(@PathVariable Integer id, @RequestBody CardDto cardDto) {
+    public ResponseEntity<CardDto> updateCard(@PathVariable Long id, @RequestBody CardDto cardDto) {
         log.debug("Updating card with ID: {} with data: {}", id, cardDto);
         CardEntity updateEntity = CardMapper.toEntity(cardDto);
         CardDto updatedCard = CardMapper.toDto(cardService.update(id, updateEntity));
@@ -65,7 +65,7 @@ public class CardController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteCard(@PathVariable Integer id) {
+    public ResponseEntity<Void> deleteCard(@PathVariable Long id) {
         log.info("Deleting card with ID: {}", id);
         cardService.delete(id);
         log.info("Card with ID: {} deleted", id);
