+++ src/main/java/com/example/rebrain/services/SetService.java
@@ -1,9 +1,7 @@
 package com.example.rebrain.services;
 
+import com.example.rebrain.entity.SetEntity;
 import com.example.rebrain.entity.*;
-import com.example.rebrain.entity.SetEntity;
-import com.example.rebrain.entity.SetEntity;
-import com.example.rebrain.entity.SetEntity;
 import com.example.rebrain.exception.ObjectNotFoundException;
 import com.example.rebrain.dto.UpdateSetDto;
 import com.example.rebrain.repositories.CardRepo;
@@ -30,7 +28,7 @@ public class SetService {
     private final CardRepo cardRepo;
 
     public SetEntity create(SetEntity setEntity) {
-        Integer userId = ThreadLocalUserIdHolder.get();
+        Long userId = Long.valueOf(ThreadLocalUserIdHolder.get());
         setEntity.setUserId(userId);
         log.info("Saving new {}", setEntity);
         return setRepo.save(setEntity);
@@ -40,36 +38,36 @@ public class SetService {
         return setRepo.findAll();
     }
 
-    public SetEntity getOne(Integer id) throws ObjectNotFoundException {
+    public SetEntity getOne(Long id) throws ObjectNotFoundException {
         return setRepo.findById(id).orElseThrow(() -> new ObjectNotFoundException("Set with ID " + id + " not found"));
     }
 
-    public List<CardEntity> getCardsBySetId(Integer setId) {
+    public List<CardEntity> getCardsBySetId(Long setId) {
         List<CardsSetsEntity> cardsSetsEntities = cardsSetsRepo.findBySetId(setId);
-        List<Integer> cardIds = cardsSetsEntities.stream()
+        List<Long> cardIds = cardsSetsEntities.stream()
                 .map(CardsSetsEntity::getCardId)
                 .collect(Collectors.toList());
         return cardRepo.findAllById(cardIds);
     }
 
-
-    public SetEntity update(Integer id, SetEntity updateEntity) throws ObjectNotFoundException {
+    public SetEntity update(Long id, SetEntity updateEntity) throws ObjectNotFoundException {
         SetEntity setEntity = getEntityById(id);
         if (updateEntity.getTitle() != null) {
             setEntity.setTitle(updateEntity.getTitle());
         }
+
         if (updateEntity.getDescription() != null) {
             setEntity.setDescription(updateEntity.getDescription());
         }
         return setRepo.save(setEntity);
     }
 
-    public void delete(Integer id) throws ObjectNotFoundException {
+    public void delete(Long id) throws ObjectNotFoundException {
         SetEntity setEntity = getEntityById(id);
         setRepo.delete(setEntity);
     }
 
-    private SetEntity getEntityById(Integer id) throws ObjectNotFoundException {
+    private SetEntity getEntityById(Long id) throws ObjectNotFoundException {
         Optional<SetEntity> setOptional = setRepo.findById(id);
         return setOptional.orElseThrow(() -> new ObjectNotFoundException("Set with ID " + id + " not found"));
     }
